Update structs-exercise.md Data type warning
I've run into an error in the automated tests, specifically regarding the Car struct. The tests seem to expect a `uint256` for the `numberOfDoors` declaration, causing a somewhat cryptic "wrong signature" error.

Given the exercises leading up to this point, it's intuitive to lean towards the smallest uint, like `uint8` given that several fellow learners at #basecamp-chat on discord and myself have faced this, and I propose the warning in my Pr to address this.