Update CONTRIBUTING.md (#312)
Fix incorrect issue links in CONTRIBUTING.md

This commit updates the 'Suggesting Enhancements' section of the CONTRIBUTING.md file to correct the issue and new issue links. Previously pointing to the base-org/base repository, these links have been corrected to direct contributors to the base-org/web repository's issue tracker. This change ensures that contributors are guided to the correct location for reporting issues or suggesting enhancements for the Base Web project.